% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stDEA.R
\name{stDEA}
\alias{stDEA}
\title{Spatio-Temporal DEA efficiency}
\usage{
stDEA(X, Y, RTS = "vrs", ORIENTATION = "out", stp = 0.01)
}
\arguments{
\item{X}{Inputs of firms to be evaluated, a K x m matrix of observations of K firms with m inputs (firm x input). In case TRANSPOSE=TRUE the input matrix is transposed to input x firm.}

\item{Y}{Outputs of firms to be evaluated, a K x n matrix of observations of K firms with n outputs (firm x input). In case TRANSPOSE=TRUE the output matrix is transposed to output x firm.}

\item{RTS}{Text string or a number defining the underlying DEA technology / returns to scale assumption.}

\item{ORIENTATION}{Input efficiency "in" (1), output efficiency "out" (2).}

\item{stp}{the STEP..}
}
\value{
The results are returned in a Farrell object with the following components. The last three components in the list are only part of the object when SLACK=TRUE.
}
\description{
Estimates a ST-DEA frontier and calculates efficiency measures a la Farrell.
}

